Moved when-done module to utils/processing
