+++ src/storage/convert.js
@@ -1,6 +1,6 @@
 import getConvertableTextures from './get-convertable-textures.js'
 import callServices from '../utils/services/call.js'
-import whenDone from '../light/when-done.js'
+import whenDone from '../utils/processing/when-done.js'
 
 function getConverter(format) {
   return function convert(storageId, options) {
